chore: move third-party lib to independent folder
